feat: jwt auth with two roles

+++ backend/cli/views.py
@@ -3,7 +3,7 @@ import re
 from django.shortcuts import render
 from rest_framework.response import Response
 from rest_framework.views import APIView
-
+from rest_framework.permissions import IsAuthenticated
 from project.models import RepoModel
 
 from . import utilities
@@ -13,53 +13,66 @@ from .serializers import CommandSerializer
 
 
 class CommandAPIView(APIView):
+    permission_classes = (IsAuthenticated,)
+
     def post(self, request):
-        serializer = CommandSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        command = serializer.validated_data["command"]
-        response = {}
 
-        create_repo_pattern = re.compile(
-            (
-                r"^create_repo (?P<name>\S+)"
-                r"( --desc (?P<description>\S*))?"
-                r"( --gitignore (?P<gitignore_template>\S*))?"
-                r"( --home (?P<homepage>\S*))?"
-                r"( --license (?P<license_template>\S*))?"
-                r"( --vis (?P<visibility>\S*))?"
+        if request.user.role == "admin":
+            serializer = CommandSerializer(data=request.data)
+            serializer.is_valid(raise_exception=True)
+            command = serializer.validated_data["command"]
+            response = {}
+
+            create_repo_pattern = re.compile(
+                (
+                    r"^create_repo (?P<name>\S+)"
+                    r"( --desc (?P<description>\S*))?"
+                    r"( --gitignore (?P<gitignore_template>\S*))?"
+                    r"( --home (?P<homepage>\S*))?"
+                    r"( --license (?P<license_template>\S*))?"
+                    r"( --vis (?P<visibility>\S*))?"
+                )
             )
-        )
 
-        if match := re.match(create_repo_pattern, command):
-            try:
-                response = utilities.create_repo(match)
-                instance = RepoModel()
-                instance.name = response["name"]
-                instance.url = response["url"]
-                instance.save()
+            if match := re.match(create_repo_pattern, command):
+                try:
+                    response = utilities.create_repo(match)
+                    instance = RepoModel()
+                    instance.name = response["name"]
+                    instance.url = response["url"]
+                    instance.save()
 
-            finally:
-                return Response(response)
+                finally:
+                    return Response(response)
+        else:
+            response = {"message":"User not authenticated"}
+            return Response(response)
 
     def delete(self, request):
-        serializer = CommandSerializer(data=request.data)
-        serializer.is_valid(raise_exception=True)
-        command = serializer.validated_data["command"]
-        response = {}
 
-        delete_repo_pattern = re.compile(r"delete_repo (?P<repo>.+)")
+        if request.user.role == "admin":
+            serializer = CommandSerializer(data=request.data)
+            serializer.is_valid(raise_exception=True)
+            command = serializer.validated_data["command"]
+            response = {}
 
-        if match := re.match(delete_repo_pattern, command):
-            response = utilities.delete_repo(match)
+            delete_repo_pattern = re.compile(r"delete_repo (?P<repo>.+)")
 
-        if response:
-            response = dict()
-            response["message"] = "Repository deleted"
-            instance = RepoModel.objects.get(name=match.group("repo"))
-            instance.delete()
+            if match := re.match(delete_repo_pattern, command):
+                response = utilities.delete_repo(match)
+
+            if response:
+                response = dict()
+                response["message"] = "Repository deleted"
+                instance = RepoModel.objects.get(name=match.group("repo"))
+                instance.delete()
+
+            else:
+                response = dict()
+                response["message"] = "Repository not found"
 
         else:
-            response = dict()
-            response["message"] = "Repository not found"
+            response = {"message":"User not authenticated"}
 
+            
         return Response(response)
